switched to jasmine and everything runs a lot faster now
