+++ jasmine-typescript-helper.js
@@ -1,0 +1,5 @@
+const { register } = require('ts-node')
+
+register({
+    project: 'tsconfig.json'
+})
