+++ jasmine.json
@@ -1,0 +1,9 @@
+{
+  "spec_dir": "out/codetl/test",
+  "spec_files": [
+    "**/*.spec.js"
+  ],
+  "helpers": [
+    "jasmine-typescript-helper.js"
+  ]
+}
