+++ package.json
@@ -3,29 +3,25 @@
   "version": "0.0.1",
   "description": "",
   "main": "out/entry.js",
+  "type": "commonjs",
   "dependencies": {
+    "@types/jasmine": "^4.0.3",
+    "@types/node": "^18.8.5",
     "@types/sarif": "^2.1.4",
     "@types/yargs": "^17.0.12",
     "antlr4ts": "^0.5.0-alpha.4",
     "codetl-parser": "file:../codetl-parser",
     "diff": "^5.1.0",
     "get-all-files": "^4.1.0",
-    "jest": "^29.1.2",
+    "jasmine": "4.4.0",
     "tempfile": "3.0.0",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.3"
+    "ts-node": "^10.7.0",
+    "typescript": "^4.8.3",
+    "yargs": "^17.6.0"
   },
   "scripts": {
     "build": "tsc -p .",
-    "start": "ts-node src/codetl/entry.ts",
-    "cli": "node --jvm --polyglot --vm.cp=../java-code-hardener/target/java-code-hardener-0.0.1.jar out/codetl/src/codetl/entry.js",
-    "test": "node --jvm --polyglot --vm.cp=../java-code-hardener/target/java-code-hardener-0.0.1.jar ./node_modules/.bin/jest --runInBand",
-    "test-only-java": "node --jvm --polyglot --vm.cp=${JAVA_CODE_HARDENER_PATH:../java-code-hardener/target/java-code-hardener-0.0.1.jar} ./node_modules/.bin/jest -i test/java.test.ts",
-    "test-except-java": "node --jvm --polyglot ./node_modules/.bin/jest -e test/java.test.ts",
-    "debug-test": "node --inspect=4117 --jvm --polyglot --vm.cp=${JAVA_CODE_HARDENER_PATH:../java-code-hardener/target/java-code-hardener-0.0.1.jar} ./node_modules/.bin/jest"
-  },
-  "exclude": [
-    "test/**/*.test.ts"
-  ]
+    "start": "node --jvm --polyglot --vm.cp=../java-code-hardener/target/java-code-hardener-0.0.1.jar out/codetl/src/codetl/entry.js",
+    "test": "node --jvm --polyglot --vm.cp=../java-code-hardener/target/java-code-hardener-0.0.1.jar node_modules/jasmine/bin/jasmine.js --config=jasmine.json"
+  }
 }
