+++ test/java.spec.ts
@@ -1,4 +1,3 @@
-import {expect, test} from '@jest/globals';
 import {fromConfiguration} from "../src/codetl/includes";
 
 import {
@@ -12,16 +11,21 @@ import {CodeTLParser} from "../../codetl-parser/src/codetl/parser";
 import {LanguageProviderResult} from "../src/codetl/codetf";
 import {readFileSync} from "fs";
 
-test('Should get all source dirs and files', () => {
+describe('Should get all source dirs and files', () => {
     const dirs = findJavaSourceDirectories("test/java_repo/")
-    expect(dirs.length).toBe(1)
-    expect(dirs[0]).toEqual("test/java_repo/src/main/java")
+    it("It finds the only src dir", () => {
+        expect(dirs.length).toBe(1)
+        expect(dirs[0]).toEqual("test/java_repo/src/main/java")
+    })
+
 
     const files = findJavaSourceFiles(dirs)
-    expect(files.length).toBe(3)
-    expect(files).toContain(resolve("test/java_repo/src/main/java/acme_1.java"))
-    expect(files).toContain(resolve("test/java_repo/src/main/java/acme_2.java"))
-    expect(files).toContain(resolve("test/java_repo/src/main/java/acme/thing/thing.java"))
+    it("It finds the expected source files in src dir", () => {
+        expect(files.length).toBe(3)
+        expect(files).toContain(resolve("test/java_repo/src/main/java/acme_1.java"))
+        expect(files).toContain(resolve("test/java_repo/src/main/java/acme_2.java"))
+        expect(files).toContain(resolve("test/java_repo/src/main/java/acme/thing/thing.java"))
+    })
 });
 
 function parseAndSetup(ruleStr : string) {
@@ -40,7 +44,7 @@ function parseAndSetup(ruleStr : string) {
     return {rule, lp, config};
 }
 
-test("It fails to process when undefined variable is referenced", () => {
+describe("It fails to process when undefined variable is referenced", () => {
     const badRuleStr: string = `
        rule pixee:java/bad-variable-definition
        match
@@ -58,10 +62,13 @@ test("It fails to process when undefined variable is referenced", () => {
         let {rule, lp, config} = parseAndSetup(badRuleStr);
         lp.process(resolve("test/java_repo"), config, [rule]);
     }
-    expect(badInvocation).toThrowError()
+
+    it("It throws an error  when executing", () => {
+        expect(badInvocation).toThrowError()
+    })
 })
 
-test("It compiles and interprets", () => {
+describe("It compiles and interprets", () => {
     const randomRuleStr : string = `
        rule pixee:java/secure-random
        match
@@ -75,23 +82,25 @@ test("It compiles and interprets", () => {
        report "upgraded prng"
     `;
 
-    let {rule, lp, config} = parseAndSetup(randomRuleStr);
-    let results : LanguageProviderResult = lp.process(resolve("test/java_repo"), config, [rule]);
-    expect(results.vendor).toEqual("openpixee")
-    expect(results.newDependencies).toHaveLength(0)
-    expect(results.language).toEqual("java")
-    expect(results.results).toHaveLength(1)
-    expect(results.results[0].path).toEqual("/src/main/java/acme_1.java")
-    expect(results.results[0].changes).toHaveLength(4)
+    it("It creates the expected results from executing on the fake repo dir", () => {
+        let {rule, lp, config} = parseAndSetup(randomRuleStr);
+        let results : LanguageProviderResult = lp.process(resolve("test/java_repo"), config, [rule]);
+        expect(results.vendor).toEqual("openpixee")
+        expect(results.newDependencies).toHaveSize(0)
+        expect(results.language).toEqual("java")
+        expect(results.results).toHaveSize(1)
+        expect(results.results[0].path).toEqual("/src/main/java/acme_1.java")
+        expect(results.results[0].changes).toHaveSize(4)
 
-    let expectedRandomFixLines : number[] = [6,9,12,16]
-    let i =0
-    for(let change of results.results[0].changes) {
-        expect(change.category).toEqual("pixee:java/secure-random")
-        expect(change.description).toEqual("upgraded prng")
-        expect(change.properties.size).toBe(0)
-        expect(change.lineNumber).toEqual(expectedRandomFixLines[i++])
-    }
+        let expectedRandomFixLines : number[] = [6,9,12,16]
+        let i =0
+        for(let change of results.results[0].changes) {
+            expect(change.category).toEqual("pixee:java/secure-random")
+            expect(change.description).toEqual("upgraded prng")
+            expect(change.properties.size).toBe(0)
+            expect(change.lineNumber).toEqual(expectedRandomFixLines[i++])
+        }
 
-    expect(results.results[0].diff.trim()).toEqual(readFileSync("test/diffs/acme_1_after.diff").toString().trim())
+        expect(results.results[0].diff.trim()).toEqual(readFileSync("test/diffs/acme_1_after.diff").toString().trim())
+    })
 })
