+++ tsconfig.json
@@ -1,11 +1,15 @@
 {
     "compilerOptions": {
-      "target": "ES6",
+      "target": "ES2020",
       "module": "commonjs",
       "esModuleInterop" : true,
       "outDir": "out",
       "sourceMap": true,
-      "strictNullChecks": true
+      "strictNullChecks": true,
+      "inlineSources": true
     },
-    "include": ["src"]
-  }
+    "include": [
+      "src/**/*.ts",
+      "test/**/*.spec.ts"
+    ]
+}
